requirements.txt
streamlit==0.86.0
Pillow==10.3.0
tensorflow==2.16.1
joblib==1.4.2
